Fix ActiveSupport deprecation warnings
Newer versions of activesupport show a warning when loading
hash/compact, since it is part of Ruby 2.5+. Most of our projects are on
newer versions than that, but just to ensure compatibility it seems
easier to just let ActiveSupport figure out what to load. I think the
performance overhead is negligible here.